+++ ccc-20200701/core/Models/ListVisitorChatMessagesRequest.cs
@@ -1,0 +1,86 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.CCC20200701.Models
+{
+    public class ListVisitorChatMessagesRequest : TeaModel {
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>cf584733-<em><b>-</b></em>-9699-cb77aa3b7aa6</para>
+        /// </summary>
+        [NameInMap("AccessChannelId")]
+        [Validation(Required=false)]
+        public string AccessChannelId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>9XYGTGWtq2wXzVikKuip_zeVGl6O4VJ-l-<em>-</em>-JPofhap4P7fAevuE=</para>
+        /// </summary>
+        [NameInMap("AccessToken")]
+        [Validation(Required=false)]
+        public string AccessToken { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>1650316799000</para>
+        /// </summary>
+        [NameInMap("EndTime")]
+        [Validation(Required=false)]
+        public string EndTime { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>ccc-test</para>
+        /// </summary>
+        [NameInMap("InstanceId")]
+        [Validation(Required=false)]
+        public string InstanceId { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>1737193352340::7463707254.EAUNIT</para>
+        /// </summary>
+        [NameInMap("NextPageToken")]
+        [Validation(Required=false)]
+        public string NextPageToken { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>100</para>
+        /// </summary>
+        [NameInMap("PageSize")]
+        [Validation(Required=false)]
+        public int? PageSize { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>DESC</para>
+        /// </summary>
+        [NameInMap("SortOrder")]
+        [Validation(Required=false)]
+        public string SortOrder { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>1647325450000</para>
+        /// </summary>
+        [NameInMap("StartTime")]
+        [Validation(Required=false)]
+        public long? StartTime { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>fcd020fe-****-1a272a174a7d</para>
+        /// </summary>
+        [NameInMap("VisitorId")]
+        [Validation(Required=false)]
+        public string VisitorId { get; set; }
+
+    }
+
+}